% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheet.R
\name{add_sheet}
\alias{add_sheet}
\title{Add new sheet to an existing workbook}
\usage{
add_sheet(data, workbook, sheet)
}
\arguments{
\item{data}{table that is going to be written in the workbook}

\item{workbook}{existing workbook}

\item{sheet}{name of the new sheet}
}
\value{
Overwrites workbook
}
\description{
Loads an existing workbook and adds the new sheet
}
\details{
If you wish to create and load new workbook at the same time
use \code{microData::new_workbook()}.
}
\examples{
# add_sheet(wb, "file.xlsx", "sheet")
}
\author{
Gutama Girja Urago
}
