% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Data dictionary}
\usage{
dictionary(catalog, filename, org = "wb", namecol = TRUE)
}
\arguments{
\item{catalog}{integer in the web link after catalog without "/"}

\item{filename}{string in the web link after data-dictionary without "/"}

\item{org}{name of an organization "wb" or "fao"}

\item{namecol}{if empty column for new name is inserted or not.
\code{default = TRUE}, then manually set new names in excel sheet}
}
\value{
a data frame
}
\description{
Extracts data dictionary from WB LSMS given catalog code and file name.
}
\examples{
# test <- dictionary(catalog = 6161, filename = "F4")
}
