% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_dict.R
\name{dt_dict}
\alias{dt_dict}
\title{Data dictionary}
\usage{
dt_dict(code, name, namecol = TRUE)
}
\arguments{
\item{code}{integer in the wb link after catalog}

\item{name}{string in the wb link after data-dictionary}

\item{namecol}{if empty column for new name is inserted or not.
\code{default = TRUE}, then manually set new names in excel sheet}
}
\value{
a data frame
}
\description{
Extracts data dictionary from WB LSMS given catalog code and file name.
}
\examples{

# library(rvest)
# library(fastverse)
# library(tidyverse)
# library(openxlsx)
# test <- scrap(code = 6161, name = "F4?file_name=sect3_hh_w5.dta", FALSE)
# wb <- loadWorkbook("test.xlsx")
# addWorksheet(wb, "here")
# writeData(wb, "here", tab)
# addWorksheet(wb, "that")
# writeData(wb, "that", ess)
# saveWorkbook(wb, "test.xlsx", overwrite = TRUE)
}
