% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_workbook.R
\name{new_workbook}
\alias{new_workbook}
\title{Create and load new workbook}
\usage{
new_workbook(filename, sheet = NULL, overwrite = TRUE)
}
\arguments{
\item{filename}{Workbook filename without .xlsx}

\item{sheet}{Name of sheet, if null \verb{Sheet 1} will be created}

\item{overwrite}{Overwrites an existing file}
}
\value{
Creates and loads workbook
}
\description{
Creates and loads new workbook with custom sheet name
}
\examples{
# new_workbook("test")
}
