% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_csv.R
\name{survey_csv}
\alias{survey_csv}
\title{Reads csv data file}
\usage{
survey_csv(data, label = NULL, sht = NULL, fct = TRUE)
}
\arguments{
\item{data}{is a csv data frame to be read}

\item{label}{an excel file containing names and labels of variables}

\item{sht}{specify sheet name if excel has more than one sheet}

\item{fct}{if you wish to convert strings in to factors}
}
\value{
cleaned data
}
\description{
Reads csv data file
}
\examples{
# survey_csv("test.csv")

}
