% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheet.R
\name{add_sheet}
\alias{add_sheet}
\title{Add sheet to an existing work book}
\usage{
add_sheet(data, wb, sht)
}
\arguments{
\item{data}{table that is going to be written in the workbook}

\item{wb}{existing workbook}

\item{sht}{name of the new sheet}
}
\value{
Overwrites workbook
}
\description{
Reads existing workbook and adds the new sheet
}
\examples{
# add_sheet(wb, "file.xlsx", "sheet")
}
